feat: refactor appointment scheduling component with improved layout and form handling
