--- src/app/components/AgendarCita/agendar-cita.component.css
@@ -1,106 +1,0 @@
-/* Layout and container styles */
-.min-h-screen {
-  min-height: 100vh;
-}
-
-.bg-gray-50 {
-  background-color: #f9fafb;
-}
-
-.py-8 {
-  padding-top: 2rem;
-  padding-bottom: 2rem;
-}
-
-.px-4 {
-  padding-left: 1rem;
-  padding-right: 1rem;
-}
-
-@media (min-width: 640px) {
-  .sm\:px-8 {
-    padding-left: 2rem;
-    padding-right: 2rem;
-  }
-}
-
-.max-w-5xl {
-  max-width: 64rem;
-}
-
-.mx-auto {
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.bg-white {
-  background-color: #ffffff;
-}
-
-.shadow {
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
-}
-
-.rounded-xl {
-  border-radius: 0.75rem;
-}
-
-.p-8 {
-  padding: 2rem;
-}
-
-/* Typography */
-.text-3xl {
-  font-size: 1.875rem;
-  line-height: 2.25rem;
-}
-
-.font-bold {
-  font-weight: 700;
-}
-
-.text-gray-800 {
-  color: #1f2937;
-}
-
-.mb-2 {
-  margin-bottom: 0.5rem;
-}
-
-.text-gray-500 {
-  color: #6b7280;
-}
-
-.font-medium {
-  font-weight: 500;
-}
-
-/* Form elements already in your CSS */
-/* Additional classes */
-.resize-none {
-  resize: none;
-}
-
-.input {
-  width: 100%;
-  padding: 0.5rem 1rem;
-  border: 1px solid #e5e7eb;
-  border-radius: 0.375rem;
-}
-
-.input:focus {
-  outline: none;
-  box-shadow: 0 0 0 2px rgba(16, 185, 129, 0.5);
-}
-
-.btn {
-  padding: 0.5rem 1rem;
-  background-color: #1f2937;
-  color: white;
-  border-radius: 0.375rem;
-  transition: background-color 200ms;
-}
-
-.btn:hover {
-  background-color: #111827;
-}
