+++ src/app/pages/agendar/agendar-page.component.css
@@ -1,15 +1,15 @@
-/* Container styles */
+/* Layout and container styles */
 .min-h-screen {
   min-height: 100vh;
 }
 
-.bg-gray-100 {
-  background-color: #f3f4f6;
+.bg-gray-50 {
+  background-color: #f9fafb;
 }
 
-.py-10 {
-  padding-top: 2.5rem;
-  padding-bottom: 2.5rem;
+.py-8 {
+  padding-top: 2rem;
+  padding-bottom: 2rem;
 }
 
 .px-4 {
@@ -17,7 +17,6 @@
   padding-right: 1rem;
 }
 
-/* Responsive padding */
 @media (min-width: 640px) {
   .sm\:px-8 {
     padding-left: 2rem;
@@ -25,9 +24,8 @@
   }
 }
 
-/* Layout styles */
-.max-w-6xl {
-  max-width: 72rem; /* 1152px */
+.max-w-5xl {
+  max-width: 64rem;
 }
 
 .mx-auto {
@@ -35,22 +33,26 @@
   margin-right: auto;
 }
 
-.mb-6 {
-  margin-bottom: 1.5rem;
+.bg-white {
+  background-color: #ffffff;
 }
 
-.mt-2 {
-  margin-top: 0.5rem;
+.shadow {
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
 }
 
-/* Text styles */
-.text-center {
-  text-align: center;
+.rounded-xl {
+  border-radius: 0.75rem;
 }
 
-.text-4xl {
-  font-size: 2.25rem;
-  line-height: 2.5rem;
+.p-8 {
+  padding: 2rem;
+}
+
+/* Typography */
+.text-3xl {
+  font-size: 1.875rem;
+  line-height: 2.25rem;
 }
 
 .font-bold {
@@ -61,6 +63,228 @@
   color: #1f2937;
 }
 
-.text-gray-600 {
-  color: #4b5563;
+.mb-2 {
+  margin-bottom: 0.5rem;
+}
+
+.text-gray-500 {
+  color: #6b7280;
+}
+
+.font-medium {
+  font-weight: 500;
+}
+
+/* Form elements already in your CSS */
+/* Additional classes */
+.resize-none {
+  resize: none;
+}
+
+.input {
+  width: 100%;
+  padding: 0.5rem 1rem;
+  border: 1px solid #e5e7eb;
+  border-radius: 0.375rem;
+}
+
+.input:focus {
+  outline: none;
+  box-shadow: 0 0 0 2px rgba(16, 185, 129, 0.5);
+}
+
+.btn {
+  padding: 0.5rem 1rem;
+  background-color: #1f2937;
+  color: white;
+  border-radius: 0.375rem;
+  transition: background-color 200ms;
+}
+
+.btn:hover {
+  background-color: #111827;
+}
+/* Fondo gris claro */
+.bg-gray-50 {
+  background-color: #f9fafb;
+}
+
+/* Sombra */
+.shadow {
+  box-shadow: 0 4px 6px -1px rgba(0,0,0,0.1), 0 2px 4px -1px rgba(0,0,0,0.06);
+}
+
+/* Bordes redondeados grandes */
+.rounded-xl {
+  border-radius: 0.75rem;
+}
+
+/* Padding */
+.p-8 { padding: 2rem; }
+.py-8 { padding-top: 2rem; padding-bottom: 2rem; }
+.px-4 { padding-left: 1rem; padding-right: 1rem; }
+.sm\:px-8 { padding-left: 2rem; padding-right: 2rem; }
+.mb-2 { margin-bottom: 0.5rem; }
+.mb-4 { margin-bottom: 1rem; }
+.mb-6 { margin-bottom: 1.5rem; }
+.mt-6 { margin-top: 1.5rem; }
+.pt-4 { padding-top: 1rem; }
+.gap-4 { gap: 1rem; }
+.grid { display: grid; }
+.max-w-5xl { max-width: 64rem; }
+.mx-auto { margin-left: auto; margin-right: auto; }
+
+/* Tipografía */
+.text-3xl { font-size: 1.875rem; line-height: 2.25rem; }
+.font-bold { font-weight: 700; }
+.text-gray-800 { color: #1f2937; }
+.text-gray-500 { color: #6b7280; }
+.text-sm { font-size: 0.875rem; }
+.font-medium { font-weight: 500; }
+
+/* Botón */
+.btn {
+  padding: 0.5rem 1rem;
+  background-color: #059669;
+  color: white;
+  border: none;
+  border-radius: 0.375rem;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background 0.2s;
+}
+.btn:hover {
+  background-color: #047857;
+}
+
+/* Input */
+.input {
+  width: 100%;
+  padding: 0.5rem 1rem;
+  border: 1px solid #e5e7eb;
+  border-radius: 0.375rem;
+  font-size: 1rem;
+}
+.input:focus {
+  outline: none;
+  box-shadow: 0 0 0 2px #34d399;
+}
+
+/* Otros */
+.resize-none { resize: none; }
+.bg-white { background: #fff; }
+.flex { display: flex; }
+.items-center { align-items: center; }
+.gap-4 { gap: 1rem; }
+.space-y-2 > * + * { margin-top: 0.5rem; }
+.text-green-100 { background: #d1fae5; }
+.text-green-800 { color: #065f46; }
+.bg-green-100 { background: #d1fae5; }
+.rounded-md { border-radius: 0.375rem; }
+.text-emerald-700 { color: #047857; }
+.text-emerald-900 { color: #065f46; }
+.bg-emerald-600 { background: #059669; }
+.bg-emerald-700 { background: #047857; }
+.text-white { color: #fff; }
+.ml-4 { margin-left: 1rem; }
+.underline { text-decoration: underline; }
+.border-t { border-top: 1px solid #e5e7eb; }
+
+/* Contenedor principal */
+.agendar-cita-container {
+  max-width: 1100px;
+  margin: 2rem auto;
+  background: #fff;
+  border-radius: 0.75rem;
+  box-shadow: 0 4px 16px rgba(0,0,0,0.08);
+  padding: 2rem;
+  display: flex;
+  flex-direction: column;
+  min-height: 80vh;
+  justify-content: center;
+}
+
+/* Parte superior: dos columnas */
+.superior {
+  display: flex;
+  gap: 2rem;
+  margin-bottom: 2rem;
+}
+
+/* Lado izquierdo y derecho */
+.lado-izquierdo, .lado-derecho {
+  flex: 1 1 0;
+  min-width: 0;
+}
+
+/* Parte inferior: formulario de cita */
+.inferior {
+  width: 100%;
+}
+
+/* Ajuste para que todo quepa sin scroll vertical en la mayoría de pantallas */
+@media (max-width: 1100px) {
+  .agendar-cita-container {
+    max-width: 98vw;
+    padding: 1rem;
+  }
+  .superior {
+    flex-direction: column;
+    gap: 1rem;
+  }
+}
+
+/* Opcional: reduce el tamaño de los inputs y botones para ahorrar espacio */
+.input, .btn, textarea {
+  font-size: 0.95rem;
+  padding: 0.4rem 0.75rem;
+}
+/* Switch tipo iOS */
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 50px;
+  height: 28px;
+}
+
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  transition: .4s;
+  border-radius: 34px;
+}
+
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 24px;
+  width: 24px;
+  left: 2px;
+  bottom: 2px;
+  background-color: white;
+  transition: .4s;
+  border-radius: 50%;
+}
+
+input:checked + .slider {
+  background-color: #34d399;
+}
+
+input:focus + .slider {
+  box-shadow: 0 0 1px #34d399;
+}
+
+input:checked + .slider:before {
+  transform: translateX(22px);
 }
