+++ src/app/pages/agendar/agendar-page.component.html
@@ -1,11 +1,110 @@
-<!-- agendar-page.component.html -->
-<div class="min-h-screen bg-gray-100 py-10 px-4 sm:px-8">
-  <div class="max-w-6xl mx-auto">
-    <div class="mb-6 text-center">
-      <h1 class="text-4xl font-bold text-gray-800">Agendar Cita</h1>
-      <p class="text-gray-600 mt-2">Completa el formulario para registrar una nueva cita</p>
+<div class="agendar-cita-container">
+  <!-- Título y descripción -->
+  <h1 class="text-3xl font-bold text-gray-800 mb-2">Agendar Cita</h1>
+  <p class="text-gray-500 mb-6">Complete el formulario para agendar una cita para su mascota</p>
+
+  <div class="superior">
+    <!-- Lado izquierdo: Cliente -->
+    <div class="lado-izquierdo">
+      <h2 class="text-xl font-semibold mb-4">Información del Cliente</h2>
+      <!-- 🔘 Switch Cliente -->
+      <div class="mb-4 flex items-center gap-4">
+        <label for="toggle-cliente" class="text-sm text-gray-700">Nuevo Cliente</label>
+        <label class="switch">
+          <input id="toggle-cliente" type="checkbox" [checked]="nuevoCliente()"
+            (change)="nuevoCliente.set(!nuevoCliente())" />
+          <span class="slider"></span>
+        </label>
+      </div>
+      <!-- 🧍‍♀️ Formulario Cliente -->
+      <div *ngIf="nuevoCliente(); else buscarCliente">
+        <form [formGroup]="clienteForm" class="grid gap-4 mb-6">
+          <input class="input" type="text" formControlName="nombre" placeholder="Nombre" />
+          <input class="input" type="text" formControlName="apellido" placeholder="Apellido" />
+          <input class="input" type="text" formControlName="direccion" placeholder="Dirección" />
+          <input class="input" type="tel" formControlName="telefono" placeholder="Teléfono" />
+          <input class="input" type="email" formControlName="email" placeholder="Correo electrónico" />
+          <input class="input" type="text" formControlName="ciudad" placeholder="Ciudad" />
+        </form>
+      </div>
+      <ng-template #buscarCliente>
+        <div class="mb-4 space-y-2">
+          <input type="text" #clienteInput placeholder="Buscar cliente por nombre" class="input" />
+          <button class="btn" type="button" (click)="buscarClientePorNombre(clienteInput.value)">Buscar</button>
+          <div *ngIf="clienteEncontrado()" class="bg-green-100 text-green-800 px-3 py-2 rounded-md text-sm">
+            <strong>Cliente encontrado:</strong>
+            {{ clienteEncontrado()?.nombre }} {{ clienteEncontrado()?.apellido }}
+            <button type="button" class="ml-4 text-sm underline text-emerald-700 hover:text-emerald-900"
+              (click)="usarClienteExistente()">
+              Usar este cliente
+            </button>
+          </div>
+        </div>
+      </ng-template>
     </div>
 
-    <app-agendar-cita></app-agendar-cita>
+    <!-- Lado derecho: Mascota -->
+    <div class="lado-derecho">
+      <h2 class="text-xl font-semibold mb-4">Información de la Mascota</h2>
+      <!-- 🔘 Switch Mascota -->
+      <div class="mb-4 flex items-center gap-4">
+        <label for="toggle-mascota" class="text-sm text-gray-700">Nueva Mascota</label>
+        <label class="switch">
+          <input id="toggle-mascota" type="checkbox" [checked]="nuevaMascota()"
+            (change)="nuevaMascota.set(!nuevaMascota())" />
+          <span class="slider"></span>
+        </label>
+      </div>
+      <!-- 🐾 Formulario Mascota -->
+      <div *ngIf="nuevaMascota(); else buscarMascota">
+        <form [formGroup]="mascotaForm" class="grid gap-4 mb-6">
+          <input class="input" type="text" formControlName="nombre" placeholder="Nombre de la mascota" />
+          <input class="input" type="number" formControlName="edad" placeholder="Edad (en años)" />
+          <select class="input" formControlName="razaId" required>
+            <option [ngValue]="null" disabled selected>Seleccione una raza</option>
+            <option *ngFor="let raza of razas$ | async" [ngValue]="raza.razaId">
+              {{ raza.nombre }}
+            </option>
+          </select>
+        </form>
+      </div>
+      <ng-template #buscarMascota>
+        <div class="mb-4 space-y-2">
+          <input type="text" #mascotaInput placeholder="Buscar mascota por nombre" class="input" />
+          <button class="btn" type="button" (click)="buscarMascotaPorNombre(mascotaInput.value)">Buscar</button>
+          <div *ngIf="mascotaEncontrada()" class="bg-green-100 text-green-800 px-3 py-2 rounded-md text-sm">
+            <strong>Mascota encontrada:</strong>
+            {{ mascotaEncontrada()?.nombre }}
+            <button type="button" class="ml-4 text-sm underline text-emerald-700 hover:text-emerald-900"
+              (click)="usarMascotaExistente()">
+              Usar esta mascota
+            </button>
+          </div>
+        </div>
+      </ng-template>
+    </div>
+  </div>
+
+  <!-- Inferior: Formulario de Cita -->
+  <div class="inferior">
+    <h2 class="text-xl font-semibold mb-4">Información de la Cita</h2>
+    <!-- 📅 Formulario de Cita -->
+    <form [formGroup]="citaForm" class="grid gap-4 border-t pt-4">
+      <div class="grid grid-cols-2 gap-4">
+        <input class="input" type="date" formControlName="fecha" />
+        <input class="input" type="time" formControlName="hora" />
+      </div>
+      <select class="input" formControlName="tipoServicioId" required>
+        <option [ngValue]="null" disabled selected>Seleccione un tipo de servicio</option>
+        <option *ngFor="let tipo of tiposServicio$ | async" [ngValue]="tipo.tipoServicioId">
+          {{ tipo.nombre }}
+        </option>
+      </select>
+      <textarea class="input resize-none" rows="3" formControlName="motivo" placeholder="Motivo de la cita"></textarea>
+      <button type="button" class="btn bg-emerald-600 hover:bg-emerald-700 text-white font-medium" (click)="agendar()"
+        [disabled]="isLoading()">
+        {{ isLoading() ? 'Agendando...' : 'Agendar Cita' }}
+      </button>
+    </form>
   </div>
 </div>
