+++ src/app/pages/agendar/agendar-page.component.ts
@@ -1,12 +1,154 @@
-import { Component } from '@angular/core';
+// agendar-cita.component.ts
+import { Component, computed, signal } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { ClienteService } from '../../core/services/cliente.service';
+import { MascotaService } from '../../core/services/mascota.service';
+import { CitaService } from '../../core/services/cita.service';
 import { CommonModule } from '@angular/common';
-import { AgendarCitaComponent } from '../../components/AgendarCita/agendar-cita.component';
+import { switchMap, Observable, of, map } from 'rxjs';
+import { Cliente } from '../../components/shared/interfaces/cliente.model';
+import { Mascota } from '../../components/shared/interfaces/mascota.model';
+import { Cita } from '../../components/shared/interfaces/cita.model';
+import { Raza } from '../../components/shared/interfaces/Raza.model';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 
 @Component({
-  selector: 'app-agendar-page',
-  standalone: true,
-  imports: [CommonModule, AgendarCitaComponent],
+  selector: 'app-agendar-cita',
   templateUrl: './agendar-page.component.html',
-  styleUrls: ['./agendar-page.component.css']
+  styleUrls: ['./agendar-page.component.css'],
+  standalone: true,
+  imports: [CommonModule, ReactiveFormsModule, FormsModule]
 })
-export class AgendarPageComponent {}
+export class AgendarPageComponent {
+  isLoading = signal(false);
+  nuevoCliente = signal(true);
+  nuevaMascota = signal(true);
+
+  clienteEncontrado = signal<Cliente | null>(null);
+  mascotaEncontrada = signal<Mascota | null>(null);
+
+  clienteForm: FormGroup;
+  mascotaForm: FormGroup;
+  citaForm: FormGroup;
+
+  razas$: Observable<Raza[]>;
+
+  tiposServicio$: Observable<{ tipoServicioId: number; nombre: string }[]> = of([]);
+
+  constructor(
+    private fb: FormBuilder,
+    private clienteService: ClienteService,
+    private mascotaService: MascotaService,
+    private citaService: CitaService
+  ) {
+    this.razas$ = this.mascotaService.listarRazas().pipe(
+      map(res => res.data) // Aquí está bien
+    );
+
+    this.tiposServicio$ = this.citaService.listarTiposServicio().pipe(
+      map(res => res.data)
+    );
+
+    this.clienteForm = this.fb.group({
+      nombre: ['', Validators.required],
+      apellido: [''],
+      direccion: [''],
+      telefono: [''],
+      email: [''],
+      ciudad: ['']
+    });
+
+    this.mascotaForm = this.fb.group({
+      nombre: ['', Validators.required],
+      edad: [null, [Validators.required, Validators.min(0)]],
+      estado: ['VIVO', Validators.required],
+      razaId: [null, Validators.required], // ✅
+      clienteId: ['']
+    });
+
+    this.citaForm = this.fb.group({
+      fecha: ['', Validators.required],
+      hora: ['', Validators.required],
+      motivo: ['', Validators.required],
+      tipoServicioId: [null, Validators.required],
+    });
+  }
+
+  buscarClientePorNombre(nombre: string) {
+    this.clienteService.buscarPorNombre(nombre).subscribe(res => {
+      this.clienteEncontrado.set(res.data?.[0] || null);
+    });
+  }
+
+  buscarMascotaPorNombre(nombre: string) {
+    this.mascotaService.buscarPorNombre(nombre).subscribe(res => {
+      this.mascotaEncontrada.set(res.data || null);
+    });
+  }
+
+  usarMascotaExistente() {
+    this.nuevaMascota.set(false); // para ocultar formulario nuevo
+  }
+
+
+  usarClienteExistente() {
+    this.nuevoCliente.set(false); // para ocultar formulario nuevo
+  }
+
+
+
+
+  agendar() {
+    if (!this.nuevoCliente() && !this.clienteEncontrado()) {
+      alert('Seleccione un cliente existente');
+      return;
+    }
+
+    if (!this.nuevaMascota() && !this.mascotaEncontrada()) {
+      alert('Seleccione una mascota existente');
+      return;
+    }
+
+    this.isLoading.set(true);
+
+    const crearCliente$ = this.nuevoCliente()
+      ? this.clienteService.crear(this.clienteForm.value)
+      : of({ data: this.clienteEncontrado()! });
+
+    crearCliente$.pipe(
+      switchMap(clienteRes => {
+        const clienteId = clienteRes.data.clienteId ?? 0;
+        const mascotaData: Mascota = { ...this.mascotaForm.value, clienteId } as Mascota;
+
+        const crearMascota$ = this.nuevaMascota()
+          ? this.mascotaService.crear(mascotaData)
+          : of({ data: this.mascotaEncontrada()! });
+
+        return crearMascota$.pipe(
+          switchMap(mascotaRes => {
+            // Ajusta estos valores según tu lógica o formulario
+            const citaPayload: Cita = {
+              fechaRegistro: new Date().toISOString(),
+              tipoServicioId: 1, // Cambia según tu lógica o selecciona en el formulario
+              mascotaId: mascotaRes.data.mascotaId ?? 0, // Provide a fallback value in case mascotaId is undefined
+              clienteId: clienteId,
+              veterinarioId: 1, // Cambia según tu lógica o selecciona en el formulario
+              motivo: this.citaForm.value.motivo
+            };
+            return this.citaService.agendar(citaPayload);
+          })
+        );
+      })
+    ).subscribe({
+      next: () => {
+        this.isLoading.set(false);
+        alert('Cita registrada con éxito');
+      },
+      error: err => {
+        this.isLoading.set(false);
+        alert('Error al agendar cita');
+        console.error(err);
+      }
+    });
+  }
+}
